shorten simple state msg

+++ controller_driver/SimpleMessage.h
@@ -193,13 +193,13 @@ typedef enum
 
 struct _SmBodyRobotStatus		// ROS_MSG_ROBOT_STATUS = 13
 {
-	int drives_powered;			// Servo Power: -1=Unknown, 1=ON, 0=OFF
-	int e_stopped;				// Controller E-Stop state: -1=Unknown, 1=True(ON), 0=False(OFF)
+	INT8 drives_powered;		// Servo Power: -1=Unknown, 1=ON, 0=OFF
+	INT8 e_stopped;				// Controller E-Stop state: -1=Unknown, 1=True(ON), 0=False(OFF)
 	int error_code;				// Alarm code
-	int in_error;				// Is there an alarm:   -1=Unknown, 1=True, 0=False 
-	int in_motion;				// Is currently executing a motion command:  -1=Unknown, 1=True, 0=False 
-	int mode;  					// Controller/Pendant mode: -1=Unknown, 1=Manual(TEACH), 2=Auto(PLAY)
-	int motion_possible;		// Is the controller ready to receive motion: -1=Unknown, 1=ENABLED, 0=DISABLED 
+	INT8 in_error;				// Is there an alarm:   -1=Unknown, 1=True, 0=False 
+	INT8 in_motion;				// Is currently executing a motion command:  -1=Unknown, 1=True, 0=False 
+	INT8 mode;  				// Controller/Pendant mode: -1=Unknown, 1=Manual(TEACH), 2=Auto(PLAY), 0=Remote
+	INT8 motion_possible;		// Is the controller ready to receive motion: -1=Unknown, 1=ENABLED, 0=DISABLED 
 	UINT8 input_direct_in;		// Direct in inputs
 } __attribute__((__packed__));
 typedef struct _SmBodyRobotStatus SmBodyRobotStatus;
